Move current GPS/AIS sentences to seperate file
