+++ n2kd/gps_ais.h
@@ -1,0 +1,5 @@
+#include "common.h"
+
+extern void nmea0183VTG( StringBuffer * msg183, int src, const char * msg );
+extern void nmea0183GSA( StringBuffer * msg183, int src, const char * msg );
+extern void nmea0183GLL( StringBuffer * msg183, int src, const char * msg );
